Workshop_04/requirements.txt
langchain>=0.1.0
langchain_community>=0.0.10
langchain_openai>=0.0.5
langchain_core>=0.1.0
langchain_experimental>=0.0.50
streamlit==1.37.1
pinecone>=6.0.0
openai>=1.10.0
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.0.0
chardet>=5.0.0
uuid>=1.30